NodeLib: move the module type out of ConfigStore into NodeLib::ModuleType
ConfigStore::Module becomes a standalone NodeLib::ModuleType enum in
EModuleType.h, used by ConfigStore, NodeMaster, the bootloader OTA slave
and MainController's BudgetAllocator/UplinkHandler. Wire bytes and the
32-byte ConfigRecord are unchanged. Its operator<< prints an out-of-range
value (straight off an Announce) as its raw number.

+++ Software/Modules/MainController/UplinkHandler.cpp
@@ -211,7 +211,7 @@ void UplinkHandler::SendRoster()
 
         uint8_t payload[7];
         payload[0] = nodeId;
-        payload[1] = master.NodeModule(nodeId);
+        payload[1] = static_cast<uint8_t>(master.NodeModule(nodeId));
         payload[2] = bootloader ? 1 : 0;
         PackU32(&payload[3], master.NodeLastContactMs(nodeId));
 
@@ -243,7 +243,7 @@ void UplinkHandler::CheckNodePresence()
 
         const uint8_t payload[4] = {
             nodeId,
-            master.NodeModule(nodeId),
+            static_cast<uint8_t>(master.NodeModule(nodeId)),
             static_cast<uint8_t>(active ? 1 : 0),
             static_cast<uint8_t>(bootloader ? 1 : 0),
         };
